Implement RPM sensing
Detect both rising and falling edges, and ignore any edges within 4ms of the last valid edge.  This skips the extra edges from the glitchy RPM line, but limits the max RPM to ~3000.

TODO: update the ignored edge window to be a fraction of the current RPM to allow for higher RPM sense limit.